Add rotate(img, angle) builtin with Picceler affine lowering (90° increments, no clipping) (#32)
* Initial plan

* Add rotate DSL op with affine lowering and tests

Agent-Logs-Url: https://github.com/Robertkq/picceler/sessions/382fbc21-71b8-4d04-b77e-682175ce3488


* Refine integer parse errors for rotate angle handling

Agent-Logs-Url: https://github.com/Robertkq/picceler/sessions/382fbc21-71b8-4d04-b77e-682175ce3488


* Wrap rotate documentation text for readability

Agent-Logs-Url: https://github.com/Robertkq/picceler/sessions/382fbc21-71b8-4d04-b77e-682175ce3488


* Adjust rotate affine lowering to preserve full image dimensions

Agent-Logs-Url: https://github.com/Robertkq/picceler/sessions/0a0c44e7-1df8-4693-887e-50af54e395a6


* Refine rotate docs and mapping variable names

Agent-Logs-Url: https://github.com/Robertkq/picceler/sessions/0a0c44e7-1df8-4693-887e-50af54e395a6


* Fix test by registering RotateOp in dialect + declaring dependent dialects

* Potential fix for pull request finding


* Validate rotate angle domain to multiples of 90

Agent-Logs-Url: https://github.com/Robertkq/picceler/sessions/e2dd92c8-50fb-442a-987e-f96a08686a78


* Enforce rotate angle constraints with diagnostics

Agent-Logs-Url: https://github.com/Robertkq/picceler/sessions/e2dd92c8-50fb-442a-987e-f96a08686a78


* Use AP(arbitrary-precision) Int to hold constant and get S(igned)Ex(tended)Value for further checking

* Full rotation & manual validation for rotation op, works!

---------

+++ include/ast.h
@@ -3,6 +3,7 @@
 #include <vector>
 #include <string>
 #include <cstddef>
+#include <cstdint>
 #include <memory>
 
 #include "lexer.h"
@@ -53,7 +54,7 @@ struct StringNode : public ASTNode {
  * @brief AST node for numeric literals.
  */
 struct NumberNode : public ASTNode {
-  size_t value;
+  int64_t value;
   std::string toString() const override;
 };
